CX2 definitions have been moved to the model
