+++ src/components/ToolBar/DataMenu/OpenNetworkInCytoscapeMenuItem.tsx
@@ -1,4 +1,4 @@
-import { MenuItem, Tooltip, Box} from '@mui/material'
+import { MenuItem, Tooltip, Box } from '@mui/material'
 import { ReactElement } from 'react'
 import { BaseMenuProps } from '../BaseMenuProps'
 
@@ -11,14 +11,13 @@ import { useTableStore } from '../../../store/TableStore'
 import { useViewModelStore } from '../../../store/ViewModelStore'
 import { useVisualStyleStore } from '../../../store/VisualStyleStore'
 import { useNetworkSummaryStore } from '../../../store/NetworkSummaryStore'
-import { exportNetworkToCx2 } from '../../../store/exportCX'
+import { exportNetworkToCx2 } from '../../../store/io/exportCX'
 import { Network } from '../../../models/NetworkModel'
-
 
 export const OpenNetworkInCytoscapeMenuItem = (
   props: BaseMenuProps,
 ): ReactElement => {
-  const cyndex = new CyNDEx();
+  const cyndex = new CyNDEx()
   const currentNetworkId = useWorkspaceStore(
     (state) => state.workspace.currentNetworkId,
   )
@@ -39,7 +38,6 @@ export const OpenNetworkInCytoscapeMenuItem = (
     state.networks.get(currentNetworkId),
   ) as Network
 
-
   const openNetworkInCytoscape = async (): Promise<void> => {
     const cx = exportNetworkToCx2(
       network,
@@ -51,32 +49,27 @@ export const OpenNetworkInCytoscapeMenuItem = (
       `Copy of ${summary.name}`,
     )
     try {
-        await cyndex.postCX2NetworkToCytoscape(cx);
+      await cyndex.postCX2NetworkToCytoscape(cx)
     } catch (e) {
-        console.log(e)
-        console.log("Cannot find Cytoscape!")
+      console.log(e)
+      console.log('Cannot find Cytoscape!')
     }
     props.handleClose()
   }
-
 
   const handleOpenNetworkInCytoscape = async (): Promise<void> => {
     await openNetworkInCytoscape()
   }
 
-
-
   const menuItem = (
-    <MenuItem
-      onClick={handleOpenNetworkInCytoscape}
-    >
+    <MenuItem onClick={handleOpenNetworkInCytoscape}>
       Open a copy of the current network in Cytoscape
     </MenuItem>
   )
 
-    return (
-      <Tooltip title="Download and open Cytoscape to open network">
-        <Box>{menuItem}</Box>
-      </Tooltip>
-    )
-  }
+  return (
+    <Tooltip title="Download and open Cytoscape to open network">
+      <Box>{menuItem}</Box>
+    </Tooltip>
+  )
+}
