+++ src/models/CxModel/Cx2/Aspect.ts
