+++ src/models/CxModel/Cx2/CoreAspects/Attribute.ts
