+++ src/models/CxModel/Cx2/CoreAspects/AttributeDeclarations.ts
