+++ src/models/CxModel/Cx2/CoreAspects/Edge.ts
