+++ src/models/CxModel/Cx2/CoreAspects/Edges.ts
