+++ src/models/CxModel/Cx2/CoreAspects/NetworkAttributes.ts
