+++ src/models/CxModel/Cx2/CoreAspects/Node.ts
