+++ src/models/CxModel/Cx2/CoreAspects/Nodes.ts
