+++ src/models/CxModel/Cx2/CoreAspects/VisualProperties.ts
