+++ src/models/CxModel/Cx2/CxValue.ts
