+++ src/models/CxModel/Cx2Network/NetworkFunctions.ts
