+++ src/models/CxModel/NetworkSummary/NDExProperty.ts
