+++ src/models/MessageModel/index.ts
@@ -1,0 +1,1 @@
+export { Message } from './Message'
